Move `admin` modules into `crates-admin` binary
The code in these modules are only used by the `crates-admin` binary, but if e.g. the `crates_io` test suite is compiled these modules needed to be recompiled too. This is solved now by moving them into the binary, which means that they will only be compiled when that binary is compiled.
