--- src/admin/mod.rs
@@ -1,13 +1,0 @@
-pub mod default_versions;
-pub mod delete_crate;
-pub mod delete_version;
-pub mod dialoguer;
-pub mod enqueue_job;
-pub mod migrate;
-pub mod populate;
-pub mod render_readmes;
-pub mod test_pagerduty;
-pub mod transfer_crates;
-pub mod upload_index;
-pub mod verify_token;
-pub mod yank_version;
